API_PassKeeper/requirements.txt
autopep8==2.0.4
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Flask==2.2.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
gunicorn==21.2.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.7
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==0.21.1
Werkzeug==2.2.3
